removed deleted option from STATUSES and setdeleted from messagemanager.
made a small bug fix in messagesystem.

implemented message deletion in inboxUI and processing in messagemenusystem. added helper functions and corresponding docstrings in message and MessageManager

options to view read, unread and archived messages are added to display but are not implemented in the system yet. will do that next